processors/nlp/requirements.txt
kafka-python
transformers
torch
psutil
